+++ satellite/src/main/java/com/linbit/linstor/layer/bcache/BCacheLayer.java
@@ -143,14 +143,14 @@ public class BCacheLayer implements DeviceLayer
     }
 
     @Override
-    public void suspendIo(AbsRscLayerObject<Resource> rscDataRef)
+    public void suspendIo(AbsRscLayerObject<Resource> ignoredRscDataRef, boolean ignoredAsRootLayerRef)
         throws ExtCmdFailedException, StorageException
     {
         throw new StorageException(SUSPEND_IO_NOT_SUPPORTED_ERR_MSG);
     }
 
     @Override
-    public void resumeIo(AbsRscLayerObject<Resource> rscDataRef)
+    public void resumeIo(AbsRscLayerObject<Resource> ignoredRscDataRef, boolean ignoredAsRootLayerRef)
         throws ExtCmdFailedException, StorageException
     {
         throw new StorageException(SUSPEND_IO_NOT_SUPPORTED_ERR_MSG);
@@ -162,7 +162,6 @@ public class BCacheLayer implements DeviceLayer
     {
         throw new StorageException(SUSPEND_IO_NOT_SUPPORTED_ERR_MSG);
     }
-
 
     @Override
     public void processResource(