+++ satellite/src/main/java/com/linbit/linstor/layer/dmsetup/cache/CacheLayer.java
@@ -1,5 +1,9 @@
 package com.linbit.linstor.layer.dmsetup.cache;
 
+import com.linbit.ChildProcessTimeoutException;
+import com.linbit.ImplementationError;
+import com.linbit.extproc.ChildProcessHandler.TimeoutType;
+import com.linbit.extproc.ExtCmd;
 import com.linbit.extproc.ExtCmdFactory;
 import com.linbit.extproc.ExtCmdFailedException;
 import com.linbit.linstor.PriorityProps;
@@ -38,12 +42,15 @@ import com.linbit.linstor.storage.data.adapter.cache.CacheVlmData;
 import com.linbit.linstor.storage.interfaces.categories.resource.AbsRscLayerObject;
 import com.linbit.linstor.storage.interfaces.categories.resource.VlmProviderObject;
 import com.linbit.linstor.storage.kinds.DeviceLayerKind;
-import com.linbit.utils.ShellUtils;
+import com.linbit.linstor.storage.utils.Commands;
 
 import javax.inject.Inject;
 import javax.inject.Provider;
 
+import java.io.IOException;
+import java.util.Collection;
 import java.util.Collections;
+import java.util.Objects;
 import java.util.Set;
 
 public class CacheLayer implements DeviceLayer
@@ -53,6 +60,8 @@ public class CacheLayer implements DeviceLayer
 
     private static final String DFLT_FEATURE = "writeback";
     private static final String DFLT_POLICY = "smq";
+    private static final long DFLT_FLUSH_TIMEOUT_IN_MS = 3 * 60_000; // 3 minutes
+    private static final long DFLT_DMSETUP_WAIT_TIMEOUT_IN_MS = 10_000;
 
     private final ErrorReporter errorReporter;
     private final AccessContext storDriverAccCtx;
@@ -152,18 +161,210 @@ public class CacheLayer implements DeviceLayer
         return true;
     }
 
+    /**
+     * Suspending a dm-cache device is relatively easy. Forcing it to also flush the data to its origin (backing)
+     * device is quite tricky.
+     * The general procedure looks as follows:
+     * <ul>
+     *  <li><code>dmsetup reload $cache_dev --table ...</code> using <code>cleaner</code> policy. This should flush
+     *      the content to the origin device eventually, but it might take about 20-30 seconds. Since we do not want
+     *      to wait for that long, we will also need to play around with the <code>migration_threashold</code> to
+     *      speed things up.</li>
+     *  <li><code>dmsetup resume $cache_dev</code>. This is necessary for the just set <code>cleaner</code> policy to
+     *      get applied/activated</li>
+     *  <li><code>dmsetup message $cache_dev 0 migration_threshold $nr_of_sectory</code> to speed up migrating data
+     *      from the cache device to the origin</li>
+     *  <li><code>dmsetup suspend $cache_dev</code>. We will still need to wait for all dirty blocks to get written,
+     *      but we at least want the application above us to stop writing new data.</li>
+     *  <li>Wait until <code>dmsetup status $cache_dev</code> shows 0 dirty blocks. We want to avoid a busy loop which
+     *      would rapid fire <code>dmsetup status</code> commands. Instead we hope for
+     *      <code>dmsetup wait $cache_dev $last_known_event_nr</code> to wake us up in dirty blocks got migrated to
+     *      the origin device.</li>
+     * </ul>
+     *
+     * This method will block until all dirty blocks are written to the origin (backing) device.
+     * @throws IOException
+     * @throws ChildProcessTimeoutException
+     * @throws AccessDeniedException
+     */
     @Override
-    public void suspendIo(AbsRscLayerObject<Resource> rscDataRef)
-        throws ExtCmdFailedException, StorageException
+    public void suspendIo(AbsRscLayerObject<Resource> rscDataRef, boolean ignoredAsRootLayerRef)
+        throws ExtCmdFailedException, StorageException, ChildProcessTimeoutException, IOException, AccessDeniedException
     {
-        DmSetupUtils.suspendIo(errorReporter, extCmdFactory, rscDataRef, true, null);
+        // the general procedure is described in the java-doc, but we still try to change the policy and send the
+        // 'migration_threshold' message to all volumes asap and only afterwards wait for all volumes to finish
+        // with the migration, hence the two loops over the same vlmData
+        Collection<? extends VlmProviderObject<Resource>> vlmDataCol = rscDataRef.getVlmLayerObjects().values();
+        for (VlmProviderObject<Resource> vlmData : vlmDataCol)
+        {
+            CacheVlmData<Resource> cacheVlmData = (CacheVlmData<Resource>) vlmData;
+            // DmSetupUtils.suspendIo(errorReporter, extCmdFactory, cacheVlmData, true, null);4
+            long endSector = Commands.getDeviceSizeInSectors(extCmdFactory.create(), cacheVlmData.getDataDevice());
+            @Nullable String cacheDevIdentifier = cacheVlmData.getIdentifier();
+            if (cacheDevIdentifier == null)
+            {
+                throw new StorageException(
+                    "Cache device has unexpectedly no identifier. " + cacheVlmData.getDevicePath() + ", " + cacheVlmData
+                        .getDataDevice()
+                );
+            }
+            DmSetupUtils.reload(
+                extCmdFactory,
+                cacheDevIdentifier,
+                getDmSetupTable(cacheVlmData, OverridePolicy.CLEANER, endSector)
+            );
+            // for some reason "cleaner" policy only starts working when we also resume-io...
+            DmSetupUtils.suspendIo(errorReporter, extCmdFactory, cacheVlmData, false, null);
+            DmSetupUtils.message(
+                extCmdFactory,
+                cacheDevIdentifier,
+                0L,
+                new String[]
+                {
+                    "migration_threshold", Long.toString(endSector)
+                }
+            );
+
+            long start = System.currentTimeMillis();
+            long now = start;
+            long dirtyCacheBlocks = 1;
+            long flushTimeout = getFlushTimeout(cacheVlmData);
+            while (dirtyCacheBlocks > 0 && now - start < flushTimeout)
+            {
+                long lastEventNr = DmSetupUtils.getLastEventNr(extCmdFactory, cacheVlmData);
+                dirtyCacheBlocks = DmSetupUtils.getDirtyCacheBlocks(extCmdFactory, cacheDevIdentifier);
+                if (dirtyCacheBlocks > 0)
+                {
+                    ExtCmd extCmd = extCmdFactory.create();
+                    extCmd.setTimeout(TimeoutType.WAIT, DFLT_DMSETUP_WAIT_TIMEOUT_IN_MS);
+                    try
+                    {
+                        DmSetupUtils.wait(extCmd, cacheDevIdentifier, lastEventNr);
+                    }
+                    catch (ChildProcessTimeoutException ignored)
+                    {
+                    }
+                }
+
+                now = System.currentTimeMillis();
+            }
+
+            if (dirtyCacheBlocks > 0)
+            {
+                throw new StorageException(
+                    "Cache device " + cacheDevIdentifier + " did not flush its data within " + flushTimeout + "ms"
+                );
+            }
+            DmSetupUtils.suspendIo(errorReporter, extCmdFactory, cacheVlmData, true, null);
+        }
+    }
+
+    private long getFlushTimeout(CacheVlmData<Resource> cacheVlmDataRef) throws AccessDeniedException
+    {
+        long ret;
+        @Nullable String val = getPrioProps(cacheVlmDataRef.getVolume()).getProp(
+            ApiConsts.KEY_CACHE_FLUSH_TIMEOUT,
+            ApiConsts.NAMESPC_CACHE
+        );
+        if (val == null)
+        {
+            ret = DFLT_FLUSH_TIMEOUT_IN_MS;
+        }
+        else
+        {
+            ret = Long.parseLong(val);
+        }
+        return ret;
+    }
+
+    private String getDmSetupTable(CacheVlmData<Resource> cacheVlmDataRef) throws StorageException
+    {
+        return getDmSetupTable(cacheVlmDataRef, OverridePolicy.DEFAULT, null);
+    }
+
+    private String getDmSetupTable(
+        CacheVlmData<Resource> cacheVlmDataRef,
+        OverridePolicy overridePolicyRef,
+        @Nullable Long endSectorRef
+    )
+        throws StorageException
+    {
+        try
+        {
+            // TODO: properly use dmsetupPolicyArgsMap
+            // ResourceDefinition rscDfn = vlm.getResourceDefinition();
+            // ResourceGroup rscGrp = rscDfn.getResourceGroup();
+            // PriorityProps prioProps = new PriorityProps(
+            // vlm.getProps(storDriverAccCtx),
+            // vlm.getVolumeDefinition().getProps(storDriverAccCtx),
+            // rscDfn.getProps(storDriverAccCtx),
+            // rscGrp.getVolumeGroupProps(storDriverAccCtx, vlmData.getVlmNr()),
+            // rscGrp.getProps(storDriverAccCtx),
+            // localNodeProps,
+            // stltConfAccessor.getReadonlyProps()
+            // );
+            // Map<String, String> dmsetupPolicyArgsMap = prioProps.renderRelativeMap(
+            // ApiConsts.NAMESPC_CACHE_POLICY_ARGS
+            // );
+            AbsVolume<Resource> vlm = cacheVlmDataRef.getVolume();
+            String dataDevice = Objects.requireNonNull(cacheVlmDataRef.getDataDevice());
+
+            @Nullable String feature;
+            String policy;
+            switch (overridePolicyRef)
+            {
+                case DEFAULT:
+                    feature = getFeature(vlm);
+                    policy = getPolicy(vlm);
+                    break;
+                case CLEANER:
+                    feature = null;
+                    policy = "cleaner";
+                    break;
+                default:
+                    throw new ImplementationError("unexpected overridePolicy:" + overridePolicyRef);
+            }
+            return DmSetupUtils.getCacheTable(
+                0L,
+                endSectorRef == null ?
+                    Commands.getDeviceSizeInSectors(extCmdFactory.create(), dataDevice) :
+                    endSectorRef,
+                dataDevice,
+                Objects.requireNonNull(cacheVlmDataRef.getCacheDevice()),
+                Objects.requireNonNull(cacheVlmDataRef.getMetaDevice()),
+                getBlocksize(vlm),
+                feature,
+                policy
+            );
+        }
+        catch (NullPointerException npe)
+        {
+            throw new StorageException(
+                "Failed to create dmsetup table for cache device for volume:" +
+                    cacheVlmDataRef.getRscLayerObject().getSuffixedResourceName() + "/" + cacheVlmDataRef.getVlmNr(),
+                npe
+            );
+        }
+        catch (InvalidKeyException | AccessDeniedException exc)
+        {
+            throw new ImplementationError(exc);
+        }
     }
 
     @Override
-    public void resumeIo(AbsRscLayerObject<Resource> rscDataRef)
+    public void resumeIo(AbsRscLayerObject<Resource> rscDataRef, boolean ignoredAsRootLayerRef)
         throws ExtCmdFailedException, StorageException
     {
-        DmSetupUtils.suspendIo(errorReporter, extCmdFactory, rscDataRef, false, null);
+        for (VlmProviderObject<Resource> vlmData : rscDataRef.getVlmLayerObjects().values())
+        {
+            CacheVlmData<Resource> cacheVlmData = (CacheVlmData<Resource>) vlmData;
+            DmSetupUtils.reload(
+                extCmdFactory,
+                cacheVlmData.getIdentifier(),
+                getDmSetupTable(cacheVlmData, OverridePolicy.DEFAULT, null)
+            );
+            DmSetupUtils.suspendIo(errorReporter, extCmdFactory, vlmData, false, null);
+        }
     }
 
     @Override
@@ -262,50 +463,19 @@ public class CacheLayer implements DeviceLayer
                 {
                     if (!vlmData.exists())
                     {
-                        VlmProviderObject<Resource> dataChild = vlmData.getChildBySuffix(RscLayerSuffixes.SUFFIX_DATA);
-                        VlmProviderObject<Resource> cacheChild = vlmData.getChildBySuffix(
-                            RscLayerSuffixes.SUFFIX_CACHE_CACHE
+                        vlmData.setDataDevice(
+                            vlmData.getChildBySuffix(RscLayerSuffixes.SUFFIX_DATA).getDevicePath()
                         );
-                        VlmProviderObject<Resource> metaChild = vlmData.getChildBySuffix(
-                            RscLayerSuffixes.SUFFIX_CACHE_META
+                        vlmData.setCacheDevice(
+                            vlmData.getChildBySuffix(RscLayerSuffixes.SUFFIX_CACHE_CACHE).getDevicePath()
                         );
-
-                        // TODO: properly use dmsetupPolicyArgsMap
-                        // ResourceDefinition rscDfn = vlm.getResourceDefinition();
-                        // ResourceGroup rscGrp = rscDfn.getResourceGroup();
-                        // PriorityProps prioProps = new PriorityProps(
-                        // vlm.getProps(storDriverAccCtx),
-                        // vlm.getVolumeDefinition().getProps(storDriverAccCtx),
-                        // rscDfn.getProps(storDriverAccCtx),
-                        // rscGrp.getVolumeGroupProps(storDriverAccCtx, vlmData.getVlmNr()),
-                        // rscGrp.getProps(storDriverAccCtx),
-                        // localNodeProps,
-                        // stltConfAccessor.getReadonlyProps()
-                        // );
-                        // Map<String, String> dmsetupPolicyArgsMap = prioProps.renderRelativeMap(
-                        // ApiConsts.NAMESPC_CACHE_POLICY_ARGS
-                        // );
-
-                        String policyArgStr = ""; // for now
-                        int policyArgsCount = ShellUtils.shellSplit(policyArgStr).size();
-                        DmSetupUtils.createCache(
-                            extCmdFactory,
-                            vlmData.getIdentifier(),
-                            dataChild.getDevicePath(),
-                            cacheChild.getDevicePath(),
-                            metaChild.getDevicePath(),
-                            getBlocksize(vlm),
-                            getFeature(vlm),
-                            getPolicy(vlm),
-                            policyArgsCount + " " + policyArgStr
+                        vlmData.setMetaDevice(
+                            vlmData.getChildBySuffix(RscLayerSuffixes.SUFFIX_CACHE_META).getDevicePath()
                         );
 
-                        vlmData.setDevicePath(
-                            String.format(
-                                FORMAT_DEV_PATH,
-                                vlmData.getIdentifier()
-                            )
-                        );
+                        DmSetupUtils.create(extCmdFactory, vlmData.getIdentifier(), getDmSetupTable(vlmData));
+
+                        vlmData.setDevicePath(String.format(FORMAT_DEV_PATH, vlmData.getIdentifier()));
                         vlmData.setExists(true);
                         errorReporter.logDebug("Cache: device (%s) created", vlmData.getDevicePath());
                     }
@@ -446,5 +616,11 @@ public class CacheLayer implements DeviceLayer
     public DeviceLayerKind getKind()
     {
         return DeviceLayerKind.CACHE;
+    }
+
+    private enum OverridePolicy
+    {
+        DEFAULT,
+        CLEANER
     }
 }